Add task author - Create migrations - Ussing session user id on controller comments and task - Create author name method on comment and task model

+++ app/controllers/tasks_controller.rb
@@ -3,7 +3,8 @@ class TasksController < ApplicationController
   skip_before_action :set_task, only: %i[index new create destroy_all finished]
 
   def index
-    @tasks = Task.where status: :ongoing
+    @tasks = Task.where user: current_user, status: :ongoing
+
     respond_to do |format|
       format.html
       format.json { render json: @tasks, status: :ok }
@@ -13,9 +14,9 @@ class TasksController < ApplicationController
   def finished
     destroy_finished = params[:finished]
 
-    Task.where(status: :finished).delete_all if destroy_finished
+    Task.where(user: current_user, status: :finished).delete_all if destroy_finished
 
-    @tasks = Task.where status: :finished
+    @tasks = Task.where user: current_user, status: :finished
     respond_to do |format|
       format.html
       format.json { render json: @tasks, status: :ok }
@@ -32,9 +33,11 @@ class TasksController < ApplicationController
 
   def create
     @task = Task.new(task_params)
+
     if @task.save
       redirect_to @task
     else
+      pp @task.errors
       render :new, status: :unprocessable_entity
     end
   end
@@ -65,11 +68,16 @@ class TasksController < ApplicationController
 
   private
 
+  def current_user
+    session[:user_id]
+  end
+
   def set_task
-    @task = Task.find(params[:id])
+    @task = Task.find_by id: params[:id], user: current_user
   end
 
   def task_params
-    params.require(:task).permit(:title, :body, :commenter, :comment, :status)
+    params.require(:task).permit(:title, :body, :comment, :status,
+                                 :user_id).with_defaults(user_id: session[:user_id])
   end
 end
